==> code/05_figures-and-correlations.rmd <==
---
title: "Figures main manuscript"
author: "Heterogeneity in direct replications in psychology and its association with effect size"
output: 
  pdf_document: 
    latex_engine: xelatex
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, fig.width = 6, fig.asp = 1.3)
```

```{r load packages, data and source helper functions}
if(!require(readr)){install.packages("readr")}
if(!require(metafor)){install.packages("metafor")}
if(!require(dplyr)){install.packages("dplyr")}
if(!require(purrr)){install.packages("purrr")}
if(!require(ggplot2)){install.packages("ggplot2")}
if(!require(cowplot)){install.packages("cowplot")}
if(!require(boot)){install.packages("boot")}

library(readr) #To load data
library(dplyr) #For data transformation
library(purrr) #For data iteration
library(metafor) #To run meta-analyses
library(ggplot2) #Plot data
library(cowplot) #Combine several plots into one
library(boot) #Bootstrap confidence intervals for correlations

dat <- read_csv("../data/collated_summary_data.csv")

source("03_helper_functions_tables_figures.r") #Load functions to prep data for figures
```

## Figure 1

```{r Plot_tau_against_I2, fig.asp = 0.618}
dat3 <- readRDS("../data/tau_simulation_results.RDS")
names(dat3) <- sort(unique(dat$effect)) #names were alphabetized upon splitting in simulation so must sort

OR2d <- c('Allowed vs. forbidden', 'Gain vs. loss framing', 'Norm of reciprocity', 'Low vs. high category scales') #effects that were transformed from OR to SMD

#Categorize effects by type of simulation for graphing
effect_type <- distinct(select(dat, effect, effect_type)) %>% 
  mutate(effect_type = ifelse(effect %in% OR2d | effect_type == "Risk difference" | 
                                effect_type == "Raw mean difference", "MD", effect_type),
         effect_type = recode(effect_type, "d" = 'SMD',
                              "r" = "Fisher's z",
                              "q" = "Cohen's q"))



dat3 <- dat3 %>% 
  bind_rows(.id = "effect") %>% #create dataframe with identifier
  left_join(., effect_type) %>% 
  select(I2, tau, effect, effect_type) %>% 
  group_by(tau, effect, effect_type) %>% 
  summarize(I2 = mean(I2)) %>% #Take mean of I2 at each tau-level and for each effect
  ungroup()

#combined line plot gives good overview.
ggplot(dat3, aes(x = tau, y = I2, group = effect, color = effect_type, linetype = effect_type)) +
  geom_line(alpha = 0.6) +
  theme_bw() +
  scale_x_continuous(name = expression(paste("Between studies standard deviation ", tau)),
                     minor_breaks = NULL) +
  scale_y_continuous(name = expression(paste("Average ", italic(I)^2," index", sep = "")),
                     minor_breaks = NULL) +
  scale_linetype_manual(name = "Effect type", 
                        values = c( "dotdash", "dotted", "solid", "dashed"),
                        breaks = c("Fisher's z", "Cohen's q", "MD", "SMD"),
                        labels = c(expression(paste("Fisher's ", italic(z))),
                                   expression(paste("Cohen's ", italic(q))),
                                   expression(paste(italic(MD))),
                                   expression(paste(italic(SMD))))) + 
  scale_color_brewer(name = "Effect type", 
                     palette = "Dark2", 
                     breaks = c("Fisher's z", "Cohen's q", "MD", "SMD"),
                     labels = c(expression(paste("Fisher's ", italic(z))),
                                   expression(paste("Cohen's ", italic(q))),
                                   expression(paste(italic(MD))),
                                   expression(paste(italic(SMD))))) +
  coord_cartesian(ylim = c(0, 100)) +
  theme(legend.position = c(0.8, 0.4),
        legend.background = element_rect(linetype = "solid", color = "black"),
        legend.key.width = unit(0.9, "cm")) +
  guides(color = guide_legend(override.aes = list(size = 0.85)))
  

# ggsave("../figures/tau-I2.png", dpi = 600, height =  4 , width = 6.47)
# ggsave("../figures/Figure1_tau-I2.pdf", dpi = 600, height =  4 , width = 6.47, device = cairo_pdf())
```

*Figure 1.* Result of simulation relating $I^2$-values to between studies standard deviation. Each line represent one of 67 effects. Tau ($\tau$) is not directly comparable across effect size measures. Code to reproduce table: osf.io/u2t3r 

## Figure 2

```{r Density-plot-prep}
#1) I2 plot----
dens <- readRDS("../data/power_simulation_results.RDS")
names(dens) <-  sort(unique(dat$effect))

#Simulated data at various heterogeneity levels, prep for plotting
I2_dist <- dens %>% 
  bind_rows(.id = "effect") %>% 
  rename(Heterogeneity = tau_index) %>% 
  mutate(Heterogeneity = recode(Heterogeneity,
                            '1' = "Zero",
                            '2' = "Small",
                            '3' = "Medium",
                            '4' = "Large"),
         Heterogeneity = as.factor(Heterogeneity))

#Observed I2 estimates, i.e, results from main table, see table.rmd
#Effects estimated with sourced function 'est_heterogeen_smd_raw'
observed <- dat %>% 
  split(.$effect) %>%  
  map_dfr(est_heterogen_smd_raw, .id = "effect") %>% 
  select(effect, I2 = s_I2) %>%
  mutate(Heterogeneity = "Observed")
```


```{r Density-plot, fig.height=4, fig.width=6.47}
I2_plot <- ggplot(I2_dist, aes(x = I2, group = Heterogeneity)) + 
  geom_histogram(data = observed, aes(y = ..density.., x = I2, fill = Heterogeneity), bins = 100, alpha = 0.6)   +
  stat_density(geom="line", position="identity", #we use stat_density instead of geom_density to get only the density lines (and no fill)
               aes(linetype = Heterogeneity, color = Heterogeneity), size = 0.75)+ 
  theme_classic() +
  coord_cartesian(xlim = c(0, 100))  +
  scale_linetype_manual(values = c( "dotdash", "dotted", "solid", "dashed"),
                        breaks = c("Zero", "Small", "Medium", "Large"), 
                        labels = c("Zero (0%)", "Small (25%)", "Medium (50%)", "Large (75%)")) +
  scale_fill_manual(name = NULL, #drop legend title so that can place directly under other legend
                    values = "#666666", #fill color
                    breaks = "Observed",
                    labels = "     Observed") + #space before Observed to align label with other labels since this guide has a different width to the others (see last line)
  scale_color_manual(values = c("#D95F02", "#E6AB02", "#1B9E77","#E7298A"), #fill colors
                     breaks = c("Zero", "Small", "Medium", "Large"),
                     labels = c("Zero (0%)", "Small (25%)", "Medium (50%)", "Large (75%)")) +
  theme(legend.position = c(0.85, 0.71), #placement of legend
        legend.spacing = unit(-0.4, "cm"), #move the two legends close to each other to seem like one
        legend.key.width = unit(0.8, "cm"), #set default keywidth wider, so we can distinguis linetypes dashed and dotdash
        legend.background = element_blank(), #remove legend background so that we can put a shared box around the two legends
        legend.box.background = element_rect(colour = "black")) + #create the shared box
  scale_y_continuous(name = "Density") +
  scale_x_continuous(name = expression(paste(italic(I)^2," index", sep = ""))) +
  guides(linetype = guide_legend(order = 1), #place linetype/color guides above fill guide
         color  = guide_legend(order = 1),
         fill = guide_legend(keywidth = unit(0.2, "cm"), #Make fill legend thinner to look like the histogram bars
                             direction = "horizontal")) #needed in combination with above
#*****************

I2_plot
# ggsave("../figures/density-I2.png", dpi = 600, height =  4, width = 6.47)
# ggsave("../figures/Figure2_density-I2.pdf", dpi = 600, height =  4, width = 6.47, device = cairo_pdf())

```

*Figure 2.* Simulated $I^2$  densities across 68 meta-analyses for zero, small, medium, and large heterogeneity according to the definitions of Higgins (2003), and the distribution of the observed $I^2$ estimates (bars) for the 68 meta-analyses. Each simulated density consists of approximately 680,000 estimates. *MD* = Mean Difference, *SMD* = Standardized Mean Difference. Code to reproduce table: osf.io/u2t3r


## Figure 3: correlations per effect type

```{r corr_per_effect_type-prep}

#Raw mean differences to SMD
MD_standardized <- dat %>% 
  filter(outcomes1_2 == "mean _ SD") %>% 
  split(.$effect) %>% 
  map_dfr(MD_fit, .id = "effect") %>% #MD_fit function is sourced
  mutate(effect_type = "Standardized Mean Difference")

#effects that were originally transformed from OR to SMD by ML1 which we refit as OR
OR2d <- c('Allowed vs. forbidden', 'Gain vs. loss framing', 
           'Norm of reciprocity', 'Low vs. high category scales') 

#odds ratio effects in ML2 we will fit as OR
ml2_or <- c("Affect & Risk", "Trolley Dilemma 1", "Framing", "Trolley Dilemma 2") 

#Risk differences from RRR1 that can be fit as odds ratios as well
rd_effects <- c("Verbal overshadowing 1", "Verbal overshadowing 2")

log_odds_ratios <- dat %>% 
  filter(effect %in% c(OR2d, ml2_or, rd_effects)) %>% 
  split(.$effect) %>% 
  map_dfr(log_or_fit, .id = "effect") %>% #log_or_fit function is sourced
  mutate(effect_type = "Log Odds Ratio")

#combine into one dataframe for plotting
tau_per_type <- MD_standardized %>% 
  bind_rows(log_odds_ratios) %>% 
  mutate(eff_size = abs(eff_size), #need the absolute values for plotting
         tau = sqrt(tau2))


#lables
set.seed(435) #for the bootstrapping
fig3_ci <- tau_per_type %>% 
  split(.$effect_type) %>% 
  map_dfr(bootfitter, .id = "effect_type") #bootfitter is a sourced function

labels <- tau_per_type %>% #row 1:2 = tau2, row 3:4 = I2, row 5:6 = H2
  split(.$effect_type) %>% 
  map_dfr(fitter, .id = "effect_type") %>% #fitter is a sourced function
  mutate(lab = paste0("r (", c(rep(8, 3), rep(41, 3)), ") = ",
                      gsub("0\\.", "\\.", r),
                      ", 95% CI ",
                      gsub("0\\.", "\\.", fig3_ci$ci))) #put output into one variable for improved plotting


```


```{r corr_per_effect_type-plot, cache = FALSE}
#tau plot----
tau_facet <- ggplot(tau_per_type, aes(x = eff_size, y = tau)) + 
  geom_point(alpha = 0.5) +
  geom_smooth(method = "lm", se = FALSE, color = "black") + #to set alpha for the line geom_line must be used rather than geom_smooth
  geom_text(data = labels[labels$index == "tau",], aes(label = lab), x = -Inf, y = Inf, hjust = -0.06, vjust = 2, size = 3.5, fontface = "italic", alpha = .6) + #hjust and vjust can be used to annotate the same position in all facets even though the scale are different
  scale_y_continuous(name = expression(paste("Between Studies Standard Deviation ", tau))) +
  scale_x_continuous(name = "Absolute Effect Size") +
  theme_classic() +
  theme(strip.background = element_blank(), #remove box around titles
        strip.text = element_text(size = 12),
        panel.border = element_rect(fill = NA, colour = "black", #add box around plots
                                    size = 0.5, linetype = "solid")) + 
  facet_wrap(~effect_type, nrow = 2, scales = "free") #use free scales since scales are different

#I2plot----
I2_facet <- ggplot(tau_per_type, aes(x = eff_size, y = s_I2)) + 
  geom_text(data = labels[labels$index == "I2",], aes(label = lab), x = -Inf, y = Inf, hjust = -0.06, vjust = 2, size = 3.5, fontface = "italic", alpha = .6) + #hjust and vjust can be used to annotate the same position in all facets even though the scale are different
  geom_point(alpha = 0.5) +
  geom_smooth(method = "lm", se = FALSE, color = "black") + #to set alpha for the line geom_line must be used rather than geom_smooth
  scale_y_continuous(name = expression(paste(italic(I)^2, " Index"))) +
  scale_x_continuous(name = "Absolute Effect Size") +
  coord_cartesian(ylim = c(0, 100)) +
  theme_classic() +
  theme(strip.background = element_blank(), #remove box around titles
        strip.text = element_text(size = 12),
        panel.border = element_rect(fill = NA, colour = "black", #add box around plots
                                    size = 0.5, linetype = "solid")) + 
  facet_wrap(~effect_type, nrow = 2, scales = "free") #use free scales since scales are different

#H2plot----
H2_facet <- ggplot(tau_per_type, aes(x = eff_size, y = H2)) + 
  geom_point(alpha = 0.5) +
  geom_smooth(method = "lm", se = FALSE, color = "black") + #to set alpha for the line geom_line must be used rather than geom_smooth
  geom_text(data = labels[labels$index == "H2",], aes(label = lab), x = -Inf, y = Inf, hjust = -0.06, vjust = 2, size = 3.5, fontface = "italic", alpha = .6) + #hjust and vjust can be used to annotate the same position in all facets even though the scale are different
  scale_y_continuous(name = expression(paste(italic(H)^2, " Index"))) +
  scale_x_continuous(name = "Absolute Effect Size") +
  theme_classic() +
  theme(strip.background = element_blank(), #remove box around titles
        strip.text = element_text(size = 12),
        panel.border = element_rect(fill = NA, colour = "black", #add box around plots
                                    size = 0.5, linetype = "solid")) + 
  facet_wrap(~effect_type, nrow = 2, scales = "free") #use free scales since scales are different

#combined----
combo <- cowplot::plot_grid(tau_facet, I2_facet, H2_facet, labels = "AUTO")
# ggsave("../figures/effect-het.png", dpi = 600, height =  7, width = 7)
# ggsave("../figures/Figure3_effect-het.pdf", dpi = 600, height =  7, width = 7, device = cairo_pdf)
combo
```

*Figure 3.* The Pearson correlation between absolute effect size and A) $\hat{\tau}$, B) $I^2$ and C) $H^2$ respectively for 43 effects that were measured as mean or standardized mean differences and 10 effects measured as odds ratios or risk differences from 13 pre-registered multi-lab replication projects. Effects reported as mean differences were standardized and odds ratios/risk differences computed as log odds ratios. *r* = Product-moment correlation, square brackets contain 95% bootstrapped percentile confidence intervals. Code to reproduce figure: osf.io/u2t3r


```{r cor_het_effect_size, eval= FALSE}

#This section gives the correlations and values reported in the section 'Heterogeneity and effect sizes' in the main manuscript (not printed if .rmd is knit).

#bootstrap confidence intervals code contribution by Amir Abdol.

set.seed(435) #same seed as for plot

#Pearson bootstrap percentile confidence intervals for SMD and logOR
cis <- tau_per_type %>% #same as in Figure 3
  split(.$effect_type) %>% 
  map_dfr(bootfitter, .id = "effect_type") #bootfitter is a sourced function

#1.1. Standardized mean differences Pearsons----
SMDs <- tau_per_type %>% #tau_per_type object from code section two up 'cor_per_effect_type_prep'
  filter(effect_type == "Standardized Mean Difference") 

#pearson tau2
cor.test(SMDs$tau, SMDs$eff_size) 
#pearson I2
cor.test(SMDs$s_I2, SMDs$eff_size)
#pearson H2
cor.test(SMDs$H2, SMDs$eff_size) 


#1.2. log OR Pearson's----
logOR <- tau_per_type %>% #tau_per_type object from code section two up 'cor_per_effect_type_prep'
  filter(effect_type == "Log Odds Ratio") 

#pearson tau2
cor.test(logOR$tau, logOR$eff_size) 
#pearson I2
cor.test(logOR$s_I2, logOR$eff_size)
#pearson H2
cor.test(logOR$H2, logOR$eff_size) 

#1.3 Excluding anchoring effects Pearson's ---- 
dropped_anchoring <- SMDs %>%
  filter(!grepl("Anchoring", effect))

#tau
cor.test(dropped_anchoring$tau, dropped_anchoring$eff_size) 
bootfit <- boot(data=dropped_anchoring, statistic=pear_tau, R=1000, formula= ~ eff_size + tau)
boot.ci(bootfit, type=c("perc", "bca")) #percentile and adjusted percental method (BCa)

#I2
cor.test(dropped_anchoring$s_I2, dropped_anchoring$eff_size) #pearson
bootfit <- boot(data=dropped_anchoring, statistic=pear_I2, R=1000, formula= ~ eff_size + s_I2) #statistic function is sourced
boot.ci(bootfit, type=c("perc", "bca")) #percentile and adjusted percental method (BCa)

#H2
cor.test(dropped_anchoring$H2, dropped_anchoring$eff_size) #pearson
bootfit <- boot(data=dropped_anchoring, statistic=pear_H2, R=1000, formula= ~ eff_size + H2)
boot.ci(bootfit, type=c("perc", "bca"))

#1.4. Spearman's rank order correlation----

#tau
cor.test(SMDs$tau, SMDs$eff_size, method = "spearman") #spearman
options(warn=-1)
bootfit <- boot(data=SMDs, statistic=spear_tau, R=1000, formula= ~ eff_size + tau)
options(warn=0)

boot.ci(bootfit, type=c("perc", "bca"))

#I2
cor.test(SMDs$s_I2, SMDs$eff_size, method = "spearman") #spearman
options(warn=-1)
bootfit <- boot(data=SMDs, statistic=spear_I2, R=1000, formula= ~ eff_size + s_I2)
options(warn=0)

boot.ci(bootfit, type=c("perc", "bca"))

#H2

cor.test(SMDs$H2, SMDs$eff_size, method = "spearman") #spearman
options(warn=-1)
bootfit <- boot(data=SMDs, statistic=spear_H2, R=1000, formula= ~ eff_size + H2)
options(warn=0)

boot.ci(bootfit, type=c("perc", "bca"))

#2. When ES not sig. diff from zero----

##When effect size not sig diff from 0 (alpha = .05), what is average heterogeneity?
nrow(SMDs[SMDs$eff_size_p > 0.05,]) #20/40 SMDs have ES not sig. diff from zero
sum(SMDs$tau[SMDs$eff_size_p > 0.05] == 0) #13/20 of these have taus == 0

mean(SMDs$tau[SMDs$eff_size_p > 0.05]) #for tau = 0.017
mean(SMDs$s_I2[SMDs$eff_size_p > 0.05]) #I2 = 3.49
mean(SMDs$H2[SMDs$eff_size_p > 0.05]) #H2 = 0.89

SMDs_not_sig <- SMDs[SMDs$eff_size_p > 0.05,]$effect

dat %>% filter(effect %in% SMDs_not_sig) %>% #amongst the non-sig SMDS
  group_by(effect) %>% 
  summarize(K = n()) %>% #compute number of labs (K) per effect
  ungroup() %>% 
  summarize(mean_k = mean(K), #average number of K = 32.8
            median = quantile(K)['50%']) #median 22.5

#% estimates that should be non-zero under central chi-square distribution. For I2 any effect size where Q <= DF ->  I2 zero
pchisq(23, 23, lower.tail = FALSE)

#logs

logOR[logOR$eff_size_p > 0.05,] #only 1 log odds ratio not signifcant. tau = 0.13, I2 = 11.8, H2 = 1.05

```
